modified:   codar.c

+++ codar.c
@@ -2,19 +2,19 @@
 
 int main(){
     
-    int nota1, nota2, nota3;
+    float nota1, nota2, nota3;
     float media;
 
     printf("*** Programa de Cálculo de Média ***\n"); //banner pra começar o programa
 
     printf("Digite a sua primeira nota: \n");
-    scanf("%d", &nota1);
+    scanf("%f", &nota1);
 
     printf("Digite a sua segunda nota: \n");
-    scanf("%d", &nota2);
+    scanf("%f", &nota2);
 
     printf("Digite a sua terceita nota: \n");
-    scanf("%d", &nota3);
+    scanf("%f", &nota3);
 
     media = (float) (nota1 + nota2 + nota3) / 3; //conversão exlícita casting - forçando p/ que tudo seja tratado como float
